traffic_class: add a cooperative traffic class
The Cooperative traffic class is used to schedule a group of BESS
modules. In this group, each module runs as a non-preemptive module and
has to yield to a specific module as its next module.

Note: this commit has to include the CooperativeTC as a friend class in
the TC base class.