bess_samples: add a BESS script that reads packets from a pcap file
This commit adds a BESS script that uses PMDPort to read packets from a
pcap file, and generates corresponding packets. The pipeline uses a
Replicate module to duplicate all egress packets so that it can generate
packets in a loop.

Note: in order to make the pipeline effective, the commit modifies the
maximum Queue (module)size and changes the DPDK compilation config to
support libpcap.