conf-faas: update test scripts to limit the traffic rate
Prior to this commit, faas test scripts can generate traffic with a much
higher bit rate (compared to the 10Gbps NIC throughput). This is because
scripts take as input both the packet rate and the packet size.

This commit enforces a hard limit on the bit rate. Therefore, test
scripts apply a limit on the max packet size given the packet rate.