flow_lb: implement L4 load balanacer with a connection table
This commit implements a L4 load balancer. It works with two tables: one
server pool table, and one connection table.

For a new flow, it calculates the flow hash, and uses it as the index to
pick a destination server from the server pool table. It then creates a
new flow entry at the connection table. For subsequent packets of this
flow, this module modifies their destination IP address, and updates
checksum.