NFVCore: fix a bug (enqueue packets when sw_q is full)
When enqueueing packets to software queues, the exact queue can be
full or almost full. Then, the packet batch can only be enqueued
partially.

This commit fixes a memory leak bug: if NFVCore does not release the
rest of packets in the batch, then the number of free packets can keep
decreasing.