manage unassigned software queues (#6)
Sometimes, all reserved cores are used up on a single worker. Then, some software queues cannot be assigned to any of the reserved cores. In this case, the NFVCtrl will take control of these software queues, and just simply dump all packets in these queues so that the packet memory can be re-used in accepting new packets.

This patch together with the next patch add 3 functions to NFVCtrl, one for adding a sw_q, one for removing a sw_q, and the last one for dumping packets in a sw_q.