NFVCore: add a logic to split |local_queue_| at the end of each short epoch
At the end of a short epoch, the short-term optimization algorithm checks the
|local_queue_| (and other software queues as well) to determine whether a
load re-balancing is required. If the algorithm determines a re-balancing,
some of flows in |local_queue_| will be moved to other software queues owned
by a normal core.

Prior to this patch, the algorithm didn't split |local_queue_| immediately.
Instead, it applies the traffic spliting decision to new packet arrivals in
the new epoch. This, however, didn't fully resolve the high tail latency
caused by transient bursts at a normal core.

This patch does the following at the end of a short epoch:
1) it checks the queue's current length via llring_count;
2) If it is smaller than the short-term packet threshold, i.e.
|epoch_packet_thresh_|, then it does nothing;
3) If it exceeds the short-term threshold, then it calls |SplitQToSwQ|
immediately to scan all packets in the queue and split them according to the
algorithm's decision.