core/modules: implement INVISV UDP proxy on top of NAT
INVISV's routing as a service (RaaS) works on top of HTTP3/QUIC.
Therefore, it requires a multi-hop UDP proxy that sits in the cloud.

This commit starts the implementation of INVISV UDP proxy as a BESS
module. The UDP proxy is implemented on top of the default dynamic NAT
module in BESS. This commit borrows the code from NAT, and creates a new
BESS module, i.e. INVISVUDPProxy, as well as its own protobuf messages.