core/modules: add a command to allow/deny a client in INVISVUDPProxy
The INVISV UDP proxy module should by default drop all traffic. It
needs to be configure to allow or deny each individual client that
uses the routing as a service (RaaS).

This commit adds a module command to INVISVUDPProxy to register a
client and configure it to be allowed or denied by this proxy service.
It also adds a test case that covers the process of registering,
allowing, and denying a client.