core/modules: refactor INVISVUDPProxy's Init message and function
Prior to this commit, INVISVUDPProxy uses the NAT module's Init
module message and function, which does not fit what the UDP proxy
requires.

This commit refactors both the Init function message and the
function itself. For the new module message, it allows users to
specify the proxy's IP/Port and the next-hop proxy's IP/Port.